Move types back into TH.hs
